pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "quollio-core"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  {name = "quollio-dev", email = "qt.dev@quollio.com"},
]
maintainers = [
  {name = "RyoAriyama", email = "ryo.arym@gmail.com"},
  {name = "tharuta", email = "35373297+TakumiHaruta@users.noreply.github.com"},
]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "blake3==0.3.3"
    ,"dbt-core==1.7.10"
    ,"dbt-snowflake==1.7.0"
    ,"dbt-redshift==1.7.1"
    ,"dbt-databricks==1.7.1"
    ,"db-dtypes==1.2.0"
    ,"jinja2==3.1.3"
    ,"PyYAML==6.0.1"
    ,"requests==2.31.0"
    ,"pyjwt==2.8.0"
    ,"redshift-connector==2.0.915"
    ,"snowflake-connector-python==3.5.0"
    ,"databricks-sdk==0.17.0"
    ,"databricks-sql-connector==2.9.5"
    ,"sqlglot==20.8.0"
    ,"google-cloud==0.34.0"
    ,"google-cloud-bigquery==3.22.0"
    ,"google-cloud-datacatalog==3.19.0"
    ,"google-cloud-datacatalog-lineage==0.3.6"
    ,"google-api-python-client==2.131.0"
]
dynamic = ["version", "description"]

[project.urls]
Source = "https://github.com/QuollioLabs/quollio-core"
Home = "https://quollio.com"

[project.optional-dependencies]
test = [
  "black>=22.3.0"
  ,"coverage>=7.3.2"
  ,"isort>=5.10.1"
  ,"pyproject-flake8>=0.0.1-alpha.2"
  ,"pytest>=5.2"
  ,"responses>=0.25.3"
  ,"dbt-osmosis==0.12.5"
]

[tool.flit.module]
name = "quollio_core"

[tool.black]
target-version = ['py39']
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_local_folder=['config',]

[tool.flake8]
max-line-length = 120
max-complexity = 18
ignore = "E402"

[tool.hatch.version]
source = "vcs"
